// apps/web-evmos/src/graphql/types/general_types.tsx
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  ActionCoin: any;
  ActionDelegation: any;
  ActionPagination: any;
  ActionRedelegation: any;
  ActionUnbondingDelegation: any;
  _coin: any;
  _dec_coin: any;
  _text: any;
  bigint: any;
  jsonb: any;
  numeric: any;
  smallint: any;
  timestamp: any;
};
// ...
export type Community_Pool = {
  __typename?: 'community_pool';
  coins: Scalars['_dec_coin'];
  height: Scalars['bigint'];
};
// ...
export type Distribution_Params = {
  __typename?: 'distribution_params';
  height: Scalars['bigint'];
  one_row_id: Scalars['Boolean'];
  params: Scalars['jsonb'];
};
// ...
export type Evmos_Inflation_Data = {
  __typename?: 'evmos_inflation_data';
  circulating_supply: Scalars['_dec_coin'];
  height: Scalars['bigint'];
  inflation_period: Scalars['Int'];
  inflation_rate: Scalars['numeric'];
  one_row_id: Scalars['Boolean'];
  skipped_epochs: Scalars['Int'];
};
// ...
export type Evmos_Inflation_Params = {
  __typename?: 'evmos_inflation_params';
  height: Scalars['bigint'];
  one_row_id: Scalars['Boolean'];
  params: Scalars['jsonb'];
};
// ...
export type Inflation = {
  __typename?: 'inflation';
  height: Scalars['bigint'];
  value: Scalars['numeric'];
};
// ...
export type Staking_Pool = {
  __typename?: 'staking_pool';
  bonded_tokens: Scalars['String'];
  height: Scalars['bigint'];
  not_bonded_tokens: Scalars['String'];
};
// ...
export type Supply = {
  __typename?: 'supply';
  coins: Scalars['_coin'];
  height: Scalars['bigint'];
};
// ...
export type Token_Price = {
  __typename?: 'token_price';
  id: Scalars['Int'];
  market_cap: Scalars['bigint'];
  price: Scalars['numeric'];
  timestamp: Scalars['timestamp'];
  /** An object relationship */
  token_unit: Token_Unit;
  unit_name: Scalars['String'];
};
// ...
export type Token_Unit = {
  __typename?: 'token_unit';
  aliases?: Maybe<Scalars['_text']>;
  denom: Scalars['String'];
  exponent: Scalars['Int'];
  price_id?: Maybe<Scalars['String']>;
  /** An object relationship */
  token: Token;
  token_name: Scalars['String'];
  /** An object relationship */
  token_price?: Maybe<Token_Price>;
  /** An array relationship */
  token_price_histories: Array<Token_Price_History>;
  /** An aggregate relationship */
  token_price_histories_aggregate: Token_Price_History_Aggregate;
  /** An array relationship */
  token_prices: Array<Token_Price>;
  /** An aggregate relationship */
  token_prices_aggregate: Token_Price_Aggregate;
};
// ...
export type MarketDataQuery = {
  communityPool: Array<{ __typename?: 'community_pool' } & Pick<Community_Pool, 'coins'>>;
  inflation: Array<{ __typename?: 'inflation' } & Pick<Inflation, 'value'>>;
  tokenPrice: Array<
    { __typename?: 'token_price' } & Pick<Token_Price, 'price'> & {
        marketCap: Token_Price['market_cap'];
      }
  >;
  supply: Array<{ __typename?: 'supply' } & Pick<Supply, 'coins'>>;
  bondedTokens: Array<{ __typename?: 'staking_pool' } & Pick<Staking_Pool, 'bonded_tokens'>>;
  distributionParams: Array<
    { __typename?: 'distribution_params' } & Pick<Distribution_Params, 'params'>
  >;
  evmosInflationData: Array<
    { __typename?: 'evmos_inflation_data' } & Pick<
      Evmos_Inflation_Data,
      'circulating_supply' | 'inflation_rate'
    >
  >;
  evmosInflationParams: Array<
    { __typename?: 'evmos_inflation_params' } & Pick<Evmos_Inflation_Params, 'params'>
  >;
};